+++ libs/solid/src/index.ts
@@ -1,3 +1,4 @@
 export * from './button'
+export * from './checkbox'
 export * from './iconbutton'
 import './styles/baseline.css'